Improve numInstances() to check SYNC_EVENTS flag before querying instances
- Relocate TestCalendar from ical4android.impl to synctools.storage.calendar
- Add nowAlignedToSecond() utility method for consistent test timing
- Update imports in all affected test classes
